bpfman-operator fixups to pr #875
Address comments from astoycos plus a few additional cleanups.

This doesn't contain any functional changes from pr #875, but rather
optimizations and better organization.
