+++ bpfman-operator/config/crd/bases/bpfman.io_uprobeprograms.yaml
@@ -111,7 +111,7 @@ spec:
                 type: object
               containers:
                 description: Containers identifes the set of containers in which to
-                  attach the uprobe. If Containers are not specified, the uprobe will
+                  attach the uprobe. If Containers is not specified, the uprobe will
                   be attached in the bpfman-agent container.  The ContainerSelector
                   is very flexible and even allows the selection of all containers
                   in a cluster.  If an attempt is made to attach uprobes to too many
